+++ 2023_noise/immune_noise.sh
@@ -12,21 +12,25 @@ for READSMLN in 15 10 5 2 1; do
             C=$(ls bam/${CT}*_Control_*.bam | grep -v mln);
             T=$(ls bam/${CT}*_${M}_*.bam | grep -v mln);
             echo "$READSMLN $CONTROLMLN $M $CT $R $T $C";
-            samtools index $T;
-            samtools index $C;
-            T_FRACTION=$(samtools idxstats $T | cut -f3 | awk -v ct=${READSMLN}000000 'BEGIN {total=0} {total += $1} END {print ct/total}')
-            T_FRACTION=$(echo $T_FRACTION | sed 's/,/./g');
-            C_FRACTION=$(samtools idxstats $C | cut -f3 | awk -v ct=${CONTROLMLN}000000 'BEGIN {total=0} {total += $1} END {print ct/total}')
-            C_FRACTION=$(echo $C_FRACTION | sed 's/,/./g');
-            echo "$T_FRACTION $C_FRACTION";
             OUT=${T/.bam/_${READSMLN}mln_noise.bam}
             echo $OUT
-            samtools view -@ 4 -H $T > ${OUT}.sam
-            samtools view -@ 4 -s ${T_FRACTION} $T >> ${OUT}.sam
-            samtools view -@ 4 -s ${C_FRACTION} $C >> ${OUT}.sam
-            samtools view -S -b ${OUT}.sam > ${OUT}.unsorted.bam
-            samtools sort ${OUT}.unsorted.bam -o ${OUT}
-            rm ${OUT}.sam ${OUT}.unsorted.bam
+            if [[ ! -f $OUT ]]; then
+                samtools index $T;
+                samtools index $C;
+                T_FRACTION=$(samtools idxstats $T | cut -f3 | awk -v ct=${READSMLN}000000 'BEGIN {total=0} {total += $1} END {print ct/total}')
+                C_FRACTION=$(samtools idxstats $C | cut -f3 | awk -v ct=${CONTROLMLN}000000 'BEGIN {total=0} {total += $1} END {print ct/total}')
+                T_FRACTION=$(echo $T_FRACTION | sed 's/,/./g');
+                C_FRACTION=$(echo $C_FRACTION | sed 's/,/./g');
+                if [[ $(echo "$T_FRACTION>1" | bc -l) -eq 1 ]]; then echo $T; T_FRACTION=1.0; fi
+                if [[ $(echo "$C_FRACTION>1" | bc -l) -eq 1 ]]; then echo $C; C_FRACTION=1.0; fi
+                echo "$T_FRACTION $C_FRACTION"
+                samtools view -@ 4 -H $T > ${OUT}.sam
+                samtools view -@ 4 -s ${T_FRACTION} $T >> ${OUT}.sam
+                samtools view -@ 4 -s ${C_FRACTION} $C >> ${OUT}.sam
+                samtools view -S -b ${OUT}.sam > ${OUT}.unsorted.bam
+                samtools sort ${OUT}.unsorted.bam -o ${OUT}
+                rm ${OUT}.sam ${OUT}.unsorted.bam
+            fi
         done
     done
 done
@@ -40,18 +44,18 @@ echo "Peak calling"
 cd ~/data/2023_Immune
 mkdir macs2
 conda activate macs2
-for M in H3K27ac H3K4me3 H3K4me1 H3K36me3 H3K27me3; do
-    for CT in TCell BCell Monocyte; do
+for CT in TCell BCell Monocyte; do
+    for M in H3K27ac H3K4me3 H3K4me1 H3K36me3 H3K27me3; do
         C=$(ls bam/${CT}*_Control_*.bam);
         for T in $(ls bam/${CT}*_${M}_*.bam); do
             echo "$M $CT $T $C";
             TN=$(basename $T);
-            FN=${TN/.bam/};
-            if [[ -z $(find macs2/ -name "${FN}_*") ]]; then
-                macs2 callpeak -t $T -c $C -n ${FN}_q0.05 -g hs -q 0.05;
-                macs2 callpeak -t $T -c $C -n ${FN}_broad0.1 -g hs --broad --broad-cutoff 0.1;
-                mv ${FN}_q0.05* macs2/;
-                mv ${FN}_broad0.1* macs2/;
+            TU=${TN/.bam/};
+            if [[ -z $(find macs2/ -name "${TU}_*") ]]; then
+                macs2 callpeak -t $T -c $C -n ${TU}_q0.05 -g hs -q 0.05;
+                macs2 callpeak -t $T -c $C -n ${TU}_broad0.1 -g hs --broad --broad-cutoff 0.1;
+                mv ${TU}_q0.05* macs2/;
+                mv ${TU}_broad0.1* macs2/;
             fi;
         done;
     done;
@@ -59,10 +63,10 @@ done;
 
 # SICER
 conda activate sicer
-D=$(pwd)
 mkdir sicer
-for M in H3K27ac H3K4me3 H3K4me1 H3K36me3 H3K27me3; do
-    for CT in TCell BCell Monocyte; do
+D=$(pwd)
+for CT in TCell BCell Monocyte; do
+    for M in H3K27ac H3K4me3 H3K4me1 H3K36me3 H3K27me3; do
         C=$(ls bam/${CT}*_Control_*.bam);
         for T in $(ls bam/${CT}*_${M}_*.bam); do
             echo "$M $CT $R $T $C";
@@ -93,8 +97,8 @@ done;
 # SPAN
 conda activate base
 mkdir span
-for M in H3K27ac H3K4me3 H3K4me1 H3K36me3 H3K27me3; do
-    for CT in TCell BCell Monocyte; do
+for CT in TCell BCell Monocyte; do
+    for M in H3K27ac H3K4me3 H3K4me1 H3K36me3 H3K27me3; do
         C=$(ls bam/${CT}*_Control_*.bam);
         for T in $(ls bam/${CT}*_${M}_*.bam); do
             echo "$M $CT $R $T $C";
